tests: dedup RamDisk helper across ext2/block tests (#658)
Replaces 14 inline RamDisk struct + BlockDevice impl copies with the
shared `kernel/tests/common/ext2_ramdisk.rs` module (introduced for #627
in PR #657). Net -842 LoC across the test surface; behavior unchanged.

Helper extensions to cover the previously inlined call sites:
  - `RamDisk::zeroed(block_size, blocks)` for the block-layer tests
    that constructed an empty disk rather than loading the golden image.
  - `RamDisk::read_slot(offset, buf)` so `ext2_orphan_finalize.rs` can
    snapshot the backing storage without touching the previously-private
    `Mutex<Vec<u8>>` field.
  - `RamDisk::patch<F>(f)` for the in-place mutation pattern (currently
    unused by migrants but kept for the patch-hook helpers in the
    pre-existing fixtures).

Migrated files:
  block_cache_sync_fs.rs, block_writeback.rs,
  ext2_block_alloc.rs, ext2_create.rs, ext2_file_read.rs,
  ext2_file_write.rs, ext2_inode_iget.rs, ext2_link.rs,
  ext2_mount.rs, ext2_orphan_chain.rs, ext2_orphan_finalize.rs,
  ext2_rename.rs, ext2_setattr.rs, mount_dev_resolve.rs

Each migrated file uses the established include pattern:

    #[path = "common/ext2_ramdisk.rs"]
    mod ext2_ramdisk;
    use ext2_ramdisk::RamDisk;

Closes #658
