+++ kernel/tests/block_writeback.rs
@@ -36,18 +36,13 @@
 extern crate alloc;
 
 use alloc::sync::Arc;
-use alloc::vec;
-use alloc::vec::Vec;
 use core::panic::PanicInfo;
-use core::sync::atomic::{AtomicU32, Ordering};
-
-use spin::Mutex;
 
 use vibix::block::cache::{BlockCache, STATE_DIRTY};
 use vibix::block::writeback::{
     self, configured_secs, reset_configured_for_tests, DEFAULT_INTERVAL_SECS,
 };
-use vibix::block::{BlockDevice, BlockError};
+use vibix::block::BlockDevice;
 use vibix::fs::vfs::ops::{StatFs, SuperOps};
 use vibix::fs::vfs::super_block::{SbFlags, SuperBlock};
 use vibix::fs::vfs::{FsId, Inode};
@@ -105,68 +100,11 @@ fn run_tests() {
 
 /// In-memory ramdisk that counts writes so tests can distinguish
 /// "daemon-driven flush" from "sync_fs explicit call".
-struct RamDisk {
-    block_size: u32,
-    storage: Mutex<Vec<u8>>,
-    writes: AtomicU32,
-}
-
-impl RamDisk {
-    fn new(block_size: u32, blocks: usize) -> Arc<Self> {
-        let bytes = (block_size as usize) * blocks;
-        Arc::new(Self {
-            block_size,
-            storage: Mutex::new(vec![0u8; bytes]),
-            writes: AtomicU32::new(0),
-        })
-    }
-
-    fn writes(&self) -> u32 {
-        self.writes.load(Ordering::Relaxed)
-    }
-}
-
-impl BlockDevice for RamDisk {
-    fn read_at(&self, offset: u64, buf: &mut [u8]) -> Result<(), BlockError> {
-        let bs = self.block_size as u64;
-        if buf.is_empty() || (buf.len() as u64) % bs != 0 || offset % bs != 0 {
-            return Err(BlockError::BadAlign);
-        }
-        let storage = self.storage.lock();
-        let end = offset
-            .checked_add(buf.len() as u64)
-            .ok_or(BlockError::OutOfRange)?;
-        if end > storage.len() as u64 {
-            return Err(BlockError::OutOfRange);
-        }
-        let off = offset as usize;
-        buf.copy_from_slice(&storage[off..off + buf.len()]);
-        Ok(())
-    }
-    fn write_at(&self, offset: u64, buf: &[u8]) -> Result<(), BlockError> {
-        self.writes.fetch_add(1, Ordering::Relaxed);
-        let bs = self.block_size as u64;
-        if buf.is_empty() || (buf.len() as u64) % bs != 0 || offset % bs != 0 {
-            return Err(BlockError::BadAlign);
-        }
-        let mut storage = self.storage.lock();
-        let end = offset
-            .checked_add(buf.len() as u64)
-            .ok_or(BlockError::Enospc)?;
-        if end > storage.len() as u64 {
-            return Err(BlockError::Enospc);
-        }
-        let off = offset as usize;
-        storage[off..off + buf.len()].copy_from_slice(buf);
-        Ok(())
-    }
-    fn block_size(&self) -> u32 {
-        self.block_size
-    }
-    fn capacity(&self) -> u64 {
-        self.storage.lock().len() as u64
-    }
-}
+// Shared `RamDisk` — see kernel/tests/common/ext2_ramdisk.rs (issues
+// #627, #658).
+#[path = "common/ext2_ramdisk.rs"]
+mod ext2_ramdisk;
+use ext2_ramdisk::RamDisk;
 
 /// Stub `SuperOps` — the writeback daemon only calls into
 /// `SuperBlock` fields (flags / draining / sb_active), never through
@@ -204,7 +142,7 @@ fn writeback_fires_after_interval() {
     // 1 s cadence is the shortest the secs-granularity knob supports;
     // fine for a test budget of ~3 s.
 
-    let disk = RamDisk::new(512, 16);
+    let disk = RamDisk::zeroed(512, 16);
     let cache = BlockCache::new(disk.clone() as Arc<dyn BlockDevice>, 512, 8);
     let dev = cache.register_device();
     let sb = make_sb(SbFlags::default());
@@ -272,7 +210,7 @@ fn join_waits_for_daemon() {
     reset_configured_for_tests();
     writeback::set_configured_secs(1);
 
-    let disk = RamDisk::new(512, 16);
+    let disk = RamDisk::zeroed(512, 16);
     let cache = BlockCache::new(disk.clone() as Arc<dyn BlockDevice>, 512, 8);
     let dev = cache.register_device();
     let sb = make_sb(SbFlags::default());
@@ -345,7 +283,7 @@ fn no_writeback_on_ro_mount() {
     reset_configured_for_tests();
     writeback::set_configured_secs(1);
 
-    let disk = RamDisk::new(512, 8);
+    let disk = RamDisk::zeroed(512, 8);
     let cache = BlockCache::new(disk.clone() as Arc<dyn BlockDevice>, 512, 4);
     let dev = cache.register_device();
     let sb = make_sb(SbFlags::RDONLY);
@@ -365,7 +303,7 @@ fn writeback_secs_zero_disables_daemon() {
     assert!(writeback::parse_cmdline(b"writeback_secs=0"));
     assert!(writeback::is_disabled());
 
-    let disk = RamDisk::new(512, 8);
+    let disk = RamDisk::zeroed(512, 8);
     let cache = BlockCache::new(disk.clone() as Arc<dyn BlockDevice>, 512, 4);
     let dev = cache.register_device();
     let sb = make_sb(SbFlags::default());
