+++ kernel/tests/ext2_mount.rs
@@ -32,13 +32,9 @@ extern crate alloc;
 
 use alloc::sync::Arc;
 use alloc::vec;
-use alloc::vec::Vec;
 use core::panic::PanicInfo;
-use core::sync::atomic::{AtomicU32, Ordering};
 
-use spin::Mutex;
-
-use vibix::block::{BlockDevice, BlockError};
+use vibix::block::BlockDevice;
 use vibix::fs::ext2::disk::{
     Ext2SuperBlock, EXT2_ERROR_FS, EXT2_MAGIC, EXT2_SB_OFF_MAGIC, EXT2_SUPERBLOCK_SIZE,
     EXT2_VALID_FS,
@@ -128,95 +124,11 @@ fn run_tests() {
     }
 }
 
-// ---------------------------------------------------------------------------
-// RamDisk: in-memory `BlockDevice` that counts writes so tests can
-// assert "RO mount drove zero writes". Matches the `RamDisk` in
-// `block_cache_sync_fs.rs` but with a constructor that takes an initial
-// byte blob (so we can seed it with the `mkfs.ext2` image).
-// ---------------------------------------------------------------------------
-
-struct RamDisk {
-    block_size: u32,
-    storage: Mutex<Vec<u8>>,
-    writes: AtomicU32,
-}
-
-impl RamDisk {
-    fn from_image(bytes: &[u8], block_size: u32) -> Arc<Self> {
-        assert!(
-            bytes.len() % block_size as usize == 0,
-            "image size {} must be a multiple of block_size {}",
-            bytes.len(),
-            block_size,
-        );
-        Arc::new(Self {
-            block_size,
-            storage: Mutex::new(bytes.to_vec()),
-            writes: AtomicU32::new(0),
-        })
-    }
-
-    fn writes(&self) -> u32 {
-        self.writes.load(Ordering::Relaxed)
-    }
-
-    fn read_slot(&self, offset: usize, buf: &mut [u8]) {
-        let storage = self.storage.lock();
-        buf.copy_from_slice(&storage[offset..offset + buf.len()]);
-    }
-
-    fn patch<F: FnOnce(&mut [u8])>(&self, f: F) {
-        let mut storage = self.storage.lock();
-        f(&mut storage);
-    }
-}
-
-impl BlockDevice for RamDisk {
-    fn read_at(&self, offset: u64, buf: &mut [u8]) -> Result<(), BlockError> {
-        let bs = self.block_size as u64;
-        if buf.is_empty() || (buf.len() as u64) % bs != 0 || offset % bs != 0 {
-            return Err(BlockError::BadAlign);
-        }
-        let storage = self.storage.lock();
-        let end = offset
-            .checked_add(buf.len() as u64)
-            .ok_or(BlockError::OutOfRange)?;
-        if end > storage.len() as u64 {
-            return Err(BlockError::OutOfRange);
-        }
-        let off = offset as usize;
-        buf.copy_from_slice(&storage[off..off + buf.len()]);
-        Ok(())
-    }
-    fn write_at(&self, offset: u64, buf: &[u8]) -> Result<(), BlockError> {
-        // Note: the counter is bumped only on a *successful* write. A
-        // malformed/out-of-range write returns an error without
-        // touching it, so assertions like "failed mount must not write
-        // to device" can't false-positive on a bad call that the
-        // driver should never issue anyway.
-        let bs = self.block_size as u64;
-        if buf.is_empty() || (buf.len() as u64) % bs != 0 || offset % bs != 0 {
-            return Err(BlockError::BadAlign);
-        }
-        let mut storage = self.storage.lock();
-        let end = offset
-            .checked_add(buf.len() as u64)
-            .ok_or(BlockError::Enospc)?;
-        if end > storage.len() as u64 {
-            return Err(BlockError::Enospc);
-        }
-        let off = offset as usize;
-        storage[off..off + buf.len()].copy_from_slice(buf);
-        self.writes.fetch_add(1, Ordering::Relaxed);
-        Ok(())
-    }
-    fn block_size(&self) -> u32 {
-        self.block_size
-    }
-    fn capacity(&self) -> u64 {
-        self.storage.lock().len() as u64
-    }
-}
+// Shared `RamDisk` — see kernel/tests/common/ext2_ramdisk.rs (issues
+// #627, #658).
+#[path = "common/ext2_ramdisk.rs"]
+mod ext2_ramdisk;
+use ext2_ramdisk::RamDisk;
 
 // ---------------------------------------------------------------------------
 // Test helpers
